Authentication, JWT, Authorization DONE

+++ nest-event-backend/src/events/events.controller.ts
@@ -1,20 +1,17 @@
-import { Body, Controller, Delete, Get, HttpCode, Logger, NotFoundException, Param, ParseIntPipe, Patch, Post, Query, UsePipes, ValidationPipe } from "@nestjs/common";
-import { InjectRepository } from "@nestjs/typeorm";
-import { Like, MoreThan, Repository } from "typeorm";
-import { Attendee } from "./attendee.entity";
+import { Body, Controller, Delete, ForbiddenException, Get, HttpCode, Logger, NotFoundException, Param, ParseIntPipe, Patch, Post, Query, UseGuards, UsePipes, ValidationPipe } from "@nestjs/common";
 import { CreateEventDto } from "./input/create-event.dto";
-import { Event } from "./event.entity";
 import { EventsService } from "./events.service";
 import { UpdateEventDto } from "../events/input/update-event.dto";
 import { ListEvents } from "./input/list.events";
+import { CurrentUser } from "src/auth/current-user.decorator";
+import { User } from "src/auth/user.entity";
+import { AuthGuardJwt } from "src/auth/auth-guard.jwt";
 
 @Controller('events')
 export class EventsController {
     private readonly logger = new Logger(EventsController.name)
 
     constructor(
-        @InjectRepository(Event) private readonly repository: Repository<Event>,
-        @InjectRepository(Attendee) private readonly attendeeRepository: Repository<Attendee>,
         private readonly eventsService: EventsService
     ) {}
 
@@ -32,41 +29,41 @@ export class EventsController {
         return events
     }
 
-    @Get('practice')
-    async practice() {
-        return await this.repository.find({
-            select: ['id', 'when'],
-            where: [{
-                id: MoreThan(3),
-                when: MoreThan(new Date('2021-02-12T13:00:00'))
-            }, {
-                description: Like('%meet%')
-            }],
-            take: 2,
-            order: {
-                id: 'DESC'
-            }
-        })
-    }
+    // @Get('practice')
+    // async practice() {
+    //     return await this.repository.find({
+    //         select: ['id', 'when'],
+    //         where: [{
+    //             id: MoreThan(3),
+    //             when: MoreThan(new Date('2021-02-12T13:00:00'))
+    //         }, {
+    //             description: Like('%meet%')
+    //         }],
+    //         take: 2,
+    //         order: {
+    //             id: 'DESC'
+    //         }
+    //     })
+    // }
 
-    @Get('practice2')
-    async practice2() {
-        // return await this.repository.findOne(1, { relations: ['attendees'] })
-        const event = await this.repository.findOne(1, { relations: ['attendees'] })
-        // const event = new Event()
-        // event.id = 1
+    // @Get('practice2')
+    // async practice2() {
+    //     // return await this.repository.findOne(1, { relations: ['attendees'] })
+    //     const event = await this.repository.findOne(1, { relations: ['attendees'] })
+    //     // const event = new Event()
+    //     // event.id = 1
 
-        const attendee = new Attendee()
-        attendee.name = 'Using Cascade'
-        // attendee.event = event
+    //     const attendee = new Attendee()
+    //     attendee.name = 'Using Cascade'
+    //     // attendee.event = event
 
-        event.attendees.push(attendee)
+    //     event.attendees.push(attendee)
 
-        // await this.attendeeRepository.save(attendee)
-        await this.repository.save(event)
+    //     // await this.attendeeRepository.save(attendee)
+    //     await this.repository.save(event)
 
-        return event
-    }
+    //     return event
+    // }
 
     @Get(':id')
     async findOne(
@@ -82,42 +79,51 @@ export class EventsController {
     }
 
     @Post()
+    @UseGuards(AuthGuardJwt)
     async create(
-        @Body() input: CreateEventDto
+        @Body() input: CreateEventDto,
+        @CurrentUser() user: User
     ) {
-        return await this.repository.save({
-            ...input,
-            when: new Date(input.when)
-        })
+        return await this.eventsService.createEvent(input, user)
     }
 
     @Patch(':id')
+    @UseGuards(AuthGuardJwt)
     async update(
         @Param('id', ParseIntPipe) id: number,
-        @Body() input: UpdateEventDto
+        @Body() input: UpdateEventDto,
+        @CurrentUser() user: User
     ) {
-        const event = await this.repository.findOne(id)
+        const event = await this.eventsService.getEvent(id)
 
         if(!event) {
             throw new NotFoundException()
         }
 
-        return await this.repository.save({
-            ...event,
-            ...input,
-            when: input.when ? new Date(input.when) : event.when
-        })
+        if(event.organizerId !== user.id) {
+            throw new ForbiddenException(null, 'You are not authorized to change this event')
+        }
+
+        return await this.eventsService.updateEvent(event, input)
     }
 
     @Delete(':id')
+    @UseGuards(AuthGuardJwt)
     @HttpCode(204)
     async remove(
-        @Param('id', ParseIntPipe) id: number
+        @Param('id', ParseIntPipe) id: number,
+        @CurrentUser() user: User
     ) {
-        const result = await this.eventsService.deleteEvent(id)
+        const event = await this.eventsService.getEvent(id)
 
-        if (result?.affected !== 1) {
-            throw new NotFoundException
+        if(!event) {
+            throw new NotFoundException()
         }
+
+        if(event.organizerId !== user.id) {
+            throw new ForbiddenException(null, 'You are not authorized to remove this event')
+        }
+
+        return await this.eventsService.deleteEvent(id)
     }
 }
